requirements.txt
pandas
pytest
python-dotenv